requirements/base.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.10.0
gunicorn==20.1.0
markdown==3.4.3
sqlparse==0.4.3
djangorestframework==3.14.0
drf-spectacular==0.26.1
